Wheatly/python/requirements.txt
# This file lists the dependencies for the project.
openai
numpy
pandas
scikit-learn
tensorflow==2.19.0
torch
flask
requests
pyserial
pydub
pytest
matplotlib
pyyaml
pyttsx3
elevenlabs
playsound==1.2.2
pyaudio
colorama
# RPi.GPIO is for Raspberry Pi hardware and will not work in standard containers
# Commented out for Docker builds:
# RPi.GPIO
google-auth
google-auth-oauthlib
google-api-python-client
spotipy 